Revert "Reorganize tuners in packages"
This reverts commit 4cdbe610ca817597fafcf7f7e962a60b6c37d5e3.
