+++ format/src/test/scala/org/calinburloiu/music/microtonalist/format/JsonTunerPluginFormatTest.scala
@@ -17,8 +17,6 @@
 package org.calinburloiu.music.microtonalist.format
 
 import org.calinburloiu.music.microtonalist.tuner.*
-import org.calinburloiu.music.microtonalist.tuner.mpe.*
-import org.calinburloiu.music.microtonalist.tuner.mts.{MtsOctave1ByteNonRealTimeTuner, MtsOctave1ByteRealTimeTuner, MtsOctave2ByteNonRealTimeTuner, MtsOctave2ByteRealTimeTuner}
 import org.calinburloiu.music.scmidi.{MidiDeviceId, PitchBendSensitivity}
 import play.api.libs.json.*
