+++ tuner/src/main/scala/org/calinburloiu/music/microtonalist/tuner/MpeChannelAllocator.scala
@@ -14,10 +14,9 @@
  *    limitations under the License.
  */
 
-package org.calinburloiu.music.microtonalist.tuner.mpe
+package org.calinburloiu.music.microtonalist.tuner
 
-import org.calinburloiu.music.microtonalist.tuner.*
-import org.calinburloiu.music.microtonalist.tuner.mpe.MpeChannelAllocator.ChannelGroup
+import org.calinburloiu.music.microtonalist.tuner.MpeChannelAllocator.ChannelGroup
 import org.calinburloiu.music.scmidi.{MidiNote, PitchClass}
 
 import scala.collection.mutable
@@ -240,7 +239,6 @@ class MpeChannelAllocator(private val zone: MpeZoneStructure) {
    * @param expressivePitchBendCents The initial expressive pitch bend in cents for the note.
    * @param preferredChannel         An optional preferred output channel (e.g., to preserve input
    *                                 allocation in MPE input mode).
-   *
    * @return [[AllocationResult]] containing the assigned channel and any notes that were dropped.
    */
   def allocate(midiNote: MidiNote,
