+++ tuner/src/main/scala/org/calinburloiu/music/microtonalist/tuner/MpeTuner.scala
@@ -14,10 +14,9 @@
  *    limitations under the License.
  */
 
-package org.calinburloiu.music.microtonalist.tuner.mpe
+package org.calinburloiu.music.microtonalist.tuner
 
 import com.typesafe.scalalogging.StrictLogging
-import org.calinburloiu.music.microtonalist.tuner.*
 import org.calinburloiu.music.scmidi.*
 
 import javax.sound.midi.{MidiMessage, ShortMessage}
