+++ tuner/src/main/scala/org/calinburloiu/music/microtonalist/tuner/MtsMessageGenerator.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2026 Calin-Andrei Burloiu
+ * Copyright 2025 Calin-Andrei Burloiu
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -14,9 +14,8 @@
  *    limitations under the License.
  */
 
-package org.calinburloiu.music.microtonalist.tuner.mts
+package org.calinburloiu.music.microtonalist.tuner
 
-import org.calinburloiu.music.microtonalist.tuner.Tuning
 import org.calinburloiu.music.scmidi.{PitchBendSensitivity, ScPitchBendMidiMessage}
 
 import java.nio.ByteBuffer
