+++ tuner/src/test/scala/org/calinburloiu/music/microtonalist/tuner/PedalTuningChangerTest.scala
@@ -17,7 +17,6 @@
 package org.calinburloiu.music.microtonalist.tuner
 
 import org.calinburloiu.music.microtonalist.tuner.PedalTuningChanger.Cc
-import org.calinburloiu.music.microtonalist.tuner.mts.MtsMessageGenerator
 import org.calinburloiu.music.scmidi.{MidiNote, ScCcMidiMessage, ScNoteOnMidiMessage}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
